% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integer_breaks.R
\name{integer_breaks}
\alias{integer_breaks}
\title{Generate integer breaks for ggplot2 Axis}
\usage{
integer_breaks(n = 5, ...)
}
\arguments{
\item{n}{Integer. Desired number of breaks (default is 5).}

\item{...}{Additional arguments passed to \code{pretty()}.}
}
\value{
A function that takes a numeric vector and returns integer axis breaks.
}
\description{
This helper function returns a function that generates integer breaks for ggplot2 axes using \code{pretty()} and \code{floor()}.
}
\examples{
scale_x_continuous(breaks = integer_breaks())

}
