% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_logo.R
\name{load_logo}
\alias{load_logo}
\title{Load ReefCloud Logo as a Raster Graphic}
\usage{
load_logo()
}
\value{
A \code{grid::rasterGrob} object containing the ReefCloud logo.
}
\description{
This function retrieves the ReefCloud logo from the specified URL, converts it from SVG format to a raster image,
and returns it as a \code{rasterGrob} object suitable for use in plots or graphical layouts.
}
\details{
The function uses the \code{rsvg} package to render the SVG logo and the \code{grid} package to create a raster graphic object.
}
\examples{
# Example usage:
logo_grob <- load_logo()
grid::grid.draw(logo_grob)

}
